refactor: Login 에 Converter 클래스 추가 및 리팩토링
