+++ src/main/java/hobbiedo/user/auth/user/dto/request/LoginRequestDTO.java
@@ -1,9 +1,14 @@
 package hobbiedo.user.auth.user.dto.request;
 
-public interface LoginRequestDTO {
-	String getUsername();
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
 
-	String getPassword();
-
-	String getUuid();
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class LoginRequestDTO {
+	String loginId;
+	String password;
 }
