+++ src/main/java/hobbiedo/user/auth/user/presentation/AuthController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import hobbiedo.user.auth.global.api.ApiResponse;
 import hobbiedo.user.auth.global.api.code.status.SuccessStatus;
 import hobbiedo.user.auth.user.application.AuthService;
+import hobbiedo.user.auth.user.converter.LoginConverter;
 import hobbiedo.user.auth.user.vo.request.LoginRequestVO;
 import hobbiedo.user.auth.user.vo.response.LoginResponseVO;
 import lombok.RequiredArgsConstructor;
@@ -24,7 +25,7 @@ public class AuthController {
 	public ApiResponse<LoginResponseVO> loginApi(@RequestBody LoginRequestVO loginVO) {
 		return ApiResponse.onSuccess(
 				SuccessStatus.USER_INTEGRATED_LOGIN_SUCCESS.getMessage(),
-				authService.login(loginVO)
+				authService.login(LoginConverter.toRequestDTO(loginVO))
 		);
 	}
 }
